+++ lib/builtin/theme_picker/theme_picker.dart
@@ -1,2 +1,3 @@
+export 'src/appearance_control.dart';
 export 'src/extension.dart';
 export 'src/settings_view.dart';